Don't remove references to a secret when it is deleted
hal5d tracks which TLS secrets each ingress references. This handles the case in
which an ingress is created before the secret(s) it references. If/when the
referenced secret(s) are created we know they are associated with an ingress and
write them out to disk. Removing a secret should not remove its secret reference
- the secret is still referenced by the ingress until the ingress is deleted.